修改通过毫秒转化为指定字符串时，精度丢失的BUG

+++ XCDateUnit/XCDateUnit.m
@@ -119,7 +119,7 @@
 
 + (NSString *)stringFromMilliseconds:(NSString *)milliseconds format:(NSString *)fmt
 {
-    NSDate *date = [NSDate dateWithTimeIntervalSince1970:(milliseconds.floatValue * 0.001)];
+    NSDate *date = [NSDate dateWithTimeIntervalSince1970:(milliseconds.doubleValue * 0.001)];
     NSDateFormatter *dateFmt = [self cacheDateFormatter];
     [dateFmt setLocale:[[NSLocale alloc] initWithLocaleIdentifier:@"zh_CN"]];
     [dateFmt setDateFormat:fmt];
